requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pdfminer.six==20221105
pydantic==2.7.1
python-dotenv==1.0.1
